Move util functions from main.js to util.js; Export functions from util.js; import function to main.js;
