+++ js/main.js
@@ -1,3 +1,10 @@
+import {
+  getRandomPositiveInteger,
+  getRandomPositiveFloat,
+  getArrayRandomElement,
+  getArrayRandomNumberElements
+} from './util';
+
 const SIMILAR_ADDS_COUNT = 10;
 const hours = ['12:00', '13:00', '14:00'];
 const realtyTypes = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
@@ -7,27 +14,6 @@ const photos = [
   'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
   'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'
 ];
-
-const getRandomPositiveInteger = (num1, num2) => {
-  const min = Math.ceil(Math.min(Math.abs(num1), Math.abs(num2)));
-  const max = Math.floor(Math.max(Math.abs(num1), Math.abs(num2)));
-  const result = Math.random() * (max - min + 1) + min;
-  return  Math.floor(result);
-};
-
-const getRandomPositiveFloat = (num1, num2, digits = 1) => {
-  const min = Math.min(Math.abs(num1), Math.abs(num2));
-  const max = Math.max(Math.abs(num1), Math.abs(num2));
-  const result = Math.random() * (max - min) + min;
-  return +result.toFixed(digits);
-};
-
-const getArrayRandomElement = (arr) => arr[getRandomPositiveInteger(0, arr.length - 1)];
-
-const getArrayRandomNumberElements = (arr) => {
-  const shuffledArray = arr.sort(() => 0.5 - Math.random());
-  return shuffledArray.slice(0, getRandomPositiveInteger(0, arr.length - 1));
-};
 
 const similarAdds = Array.from({length: SIMILAR_ADDS_COUNT}, (value, index) => {
   const userId = `0${index + 1}`.slice(-2);
