+++ js/util.js
@@ -1,0 +1,27 @@
+const getRandomPositiveInteger = (num1, num2) => {
+  const min = Math.ceil(Math.min(Math.abs(num1), Math.abs(num2)));
+  const max = Math.floor(Math.max(Math.abs(num1), Math.abs(num2)));
+  const result = Math.random() * (max - min + 1) + min;
+  return  Math.floor(result);
+};
+
+const getRandomPositiveFloat = (num1, num2, digits = 1) => {
+  const min = Math.min(Math.abs(num1), Math.abs(num2));
+  const max = Math.max(Math.abs(num1), Math.abs(num2));
+  const result = Math.random() * (max - min) + min;
+  return +result.toFixed(digits);
+};
+
+const getArrayRandomElement = (arr) => arr[getRandomPositiveInteger(0, arr.length - 1)];
+
+const getArrayRandomNumberElements = (arr) => {
+  const shuffledArray = arr.sort(() => 0.5 - Math.random());
+  return shuffledArray.slice(0, getRandomPositiveInteger(0, arr.length - 1));
+};
+
+export {
+  getRandomPositiveInteger,
+  getRandomPositiveFloat,
+  getArrayRandomElement,
+  getArrayRandomNumberElements
+};
